refactor file dir
