+++ src/agentscope/middleware/__init__.py
@@ -2,7 +2,10 @@
 """Middleware system for AgentScope agents."""
 
 from ._base import MiddlewareBase
+from ._tracing import TracingMiddleware, setup_tracing
 
 __all__ = [
     "MiddlewareBase",
+    "TracingMiddleware",
+    "setup_tracing",
 ]
